refactor(benchmark): split monolithic index.ts into commands/ and lib/ modules
